% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial_change_test.R
\name{spatial_change_test}
\alias{spatial_change_test}
\title{Spatio-Temporal Change-Point Test}
\description{Caries out a change-point test for sequences of
functional data observed at different spatial locations.  Three tests are carried out
and three pvalues are returned.}
\usage{
spatial_change_test(DATA3, loc, varprop = 0.85, ncores = "all",
  reps_per_core = 100)
}
\arguments{
\item{DATA3}{Data for testing.  This should be an 3D array with the
first coordinate indicating the temporal location on a curve, the
second the spatial location, and the third the repitition number/time series iteration.}

\item{loc}{Two column matrix of spatial locations.}

\item{varprop}{Proportion of temporal variance explained in first two
tests (third test uses all explained variance)}

\item{ncores}{Number of cores to be used for monte carlo sampling.}

\item{reps_per_core}{Number of monte carlo samples per core.}
}
\value{
Returns three values for three change-point tests.  The first normalizes each FPC, while the second two do not.  The only difference between the second and third is that the third uses all explained variance.
}
\details{
Carries out three different change point tests as outlined in Gromenko, Kokoszka,
and Reimherr (2015).  P-values are generated using Monte-Carlo.  The total MC sample size
is given by ncores*reps_per_core.  Be warned that the code will not try to use more cores
than are available, so users should choose the ncores value appropriately.  By default, the
code will use all available cores.  This will cause a computer to be slugish in other tasks
and so the user should be careful when using the default.

Data3[i,j,k] indicates an the ith observation at the jth location and on the kth curve.  It is
important that this order is accurate.  The method assumes that the spatio-temporal
covariance is separable.  Methods such as Constantinou, Kokoszka, and Reimherr (2015) can be
used to check this assumption.  This allows the spatial and temporal covariances to be
estimated separately.  The spatial covariance is assumed to be stationary and isotropic.  A
tapered bsplines estimator is then used to estiamte it nonparametrically.  No assumptions are
made on the form of the temporal covariance function.  It is estimated by pooling appropriately
across space.

The testing procedures are all based on integrated CUSUM test statistics.  The primary difference
between the three is in how they handle the temporal covariance.  The first statistic uses
FPCA to reduce the dimension of the data and normalizes by the corresponding eigenvalues.  The
second does the same, but does not normalize by the eigenvalue.  The third procedure does not use
FPCA and does not normalize by the eigenvalues.  It can be thought of as the same as procedure 2,
but with all dimensions included.  The first procedure should be used with caution as it can
be unstable for smaller sample sizes/larger numbers of FPCs.
}

