% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpt_calcs.R
\name{scpt_calcs}
\alias{scpt_calcs}
\title{A calculation function for CPT data}
\usage{
scpt_calcs(data, sounding_info, unit_weight = 20, water_table = 6,
  nu = 0.3, schnaid_g_go = 0.3, schnaid_k = 240, atm_pressure = 100,
  Nkt = 15)
}
\arguments{
\item{data}{This is a dataframe containing the sounding information}

\item{sounding_info}{This is a dataframe containing the header info for the sounding}

\item{unit_weight}{This is a user entered value. Default = 20 g/cm^3.}

\item{water_table}{This is a user entered value for the depth to water table. Default = 6 m.}

\item{nu}{User entered poisson ratio, default = 0.3}

\item{schnaid_g_go}{Default value of Schaid method g/go ratio, default = 0.3}

\item{schnaid_k}{Schnaid (2004) method K value, default = 240}

\item{atm_pressure}{Atmospheric pressure used for calculations, default = 100}

\item{Nkt}{Mayne (2009) method default Nkt value, default = 15}
}
\description{
This function takes sounding information from a typical ConeTec investigation
}
\examples{
scpt_calcs()
}
\keyword{CPT}
