% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpt_plot.R
\name{scpt_plot}
\alias{scpt_plot}
\title{A plotting function for CPT data}
\usage{
scpt_plot(data, sounding = "sounding")
}
\arguments{
\item{data}{This is a dataframe containing the sounding information}

\item{sounding}{A string to name the sounding for file purposes}
}
\description{
This function takes sounding information from a typical ConeTec investigation
}
\examples{
scpt_plot()
}
\keyword{CPT}
